Switch to webRequest.onBeforeRequest filtering
The previous tab-based listening approach had two shortcomings:

 1. It also let us filter tab URLs *after* we had already loaded the request.
    This meant that the tracking information has already been sent to the
    original host, and this extension therefore wasn't doing anything for
    privacy (only URL aesthetics).
 2. It was possible to get into URL redirect loops when ?utm_* parameters were
    dynamically added by a site's JavaScript. (#6)

We now use webRequest.onBeforeRequest to filter requests *before* they occur.
This addresses both of the above deficiencies. New permissions are required
however.

This approach also means that this extension sits in the blocking path of all
output http and https requests, so it needs to execute quickly. In support of
that, I realized I should have been pre-compiling the regular expression, so
I've made that improvement, too.

Due to the significant behavior change, I'm also going to bump the extension
version forward to 2.0.

Closes #6